refactor(adr-061): split castxml record-entity parsing into extract/headers
parse_types()/build_record_type()/parse_record_fields()/
expand_anonymous_field()/parse_bitfield_bits(), plus the vtable/RTTI layout
walk (build_vtable/collect_virtual_methods/inherited_vtable_slots/
resolved_override_keys/vtable_slot_key), now live in
abicheck/extract/headers/castxml/records.py as free functions taking
CastxmlParserContext explicitly -- the third entity module built on
castxml's shared context, after enums.py and functions.py.

ctx.vtable_slot_root/ctx.vtable_slot_extra_roots already lived on
CastxmlParserContext from the prior functions.py slice, so this move
needed no context-shape change -- only relocating the code that reads
and mutates them. collect_virtual_methods/vtable_slot_key are the first
functions in this package to mutate shared context state rather than
only read it.

_CastxmlParser's eleven matching methods (including the @staticmethod
_parse_bitfield_bits) are now one-line delegations, so every existing
caller (including tests reading a parser's private methods directly,
e.g. parser._collect_virtual_methods(...)) keeps working unchanged.

tests/test_vtable_evidence_guard.py::test_the_producer_derivation_this_rests_on
updated to read the derivation's new home (it grepped dumper_castxml.py's
source text, which now only delegates).

Clang's records.py, and templates.py on both backends, remain open --
their record-parsing bodies differ enough in shape from castxml's that
moving both backends in one pass risked the correctness this slice's own
vtable/RTTI stakes call for; see ADR-061's Phase 5 section.

Verified: python scripts/check_architecture.py (0 errors), mypy abicheck/
(clean), ruff check/format (clean on changed files), the
castxml/clang/dumper/vtable/record-focused suite (3880 passed), the full
fast suite (33993 passed, 5 pre-existing/environment failures identical
on a clean HEAD checkout, none touching this code), the real
integration-marker castxml+clang suite including vtable/RTTI cases (70
passed), and a real dump-before/dump-after byte-for-byte snapshot
comparison (identical types/vtable output, only timestamp/git-commit
metadata differed).
